Move mail testing implementation to testing package
