+++ legacy/core/src/test/java/com/fsck/k9/message/html/HtmlConverterTest.kt
@@ -2,8 +2,8 @@ package com.fsck.k9.message.html
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
-import com.fsck.k9.mail.crlf
-import com.fsck.k9.mail.removeLineBreaks
+import com.fsck.k9.mail.testing.crlf
+import com.fsck.k9.mail.testing.removeLineBreaks
 import org.junit.Test
 
 class HtmlConverterTest {
