+++ legacy/ui/legacy/src/test/java/com/fsck/k9/message/PgpMessageBuilderTest.kt
@@ -28,21 +28,21 @@ import com.fsck.k9.autocrypt.AutocryptOperationsHelper.assertMessageHasAutocrypt
 import com.fsck.k9.mail.Address
 import com.fsck.k9.mail.BoundaryGenerator
 import com.fsck.k9.mail.MessagingException
-import com.fsck.k9.mail.assertk.asBytes
-import com.fsck.k9.mail.assertk.asText
-import com.fsck.k9.mail.assertk.body
-import com.fsck.k9.mail.assertk.bodyPart
-import com.fsck.k9.mail.assertk.bodyParts
-import com.fsck.k9.mail.assertk.contentTransferEncoding
-import com.fsck.k9.mail.assertk.contentType
-import com.fsck.k9.mail.assertk.mimeType
-import com.fsck.k9.mail.assertk.parameter
-import com.fsck.k9.mail.assertk.value
 import com.fsck.k9.mail.internet.BinaryTempFileBody
 import com.fsck.k9.mail.internet.MessageIdGenerator
 import com.fsck.k9.mail.internet.MimeMessage
 import com.fsck.k9.mail.internet.MimeMultipart
 import com.fsck.k9.mail.internet.TextBody
+import com.fsck.k9.mail.testing.assertk.asBytes
+import com.fsck.k9.mail.testing.assertk.asText
+import com.fsck.k9.mail.testing.assertk.body
+import com.fsck.k9.mail.testing.assertk.bodyPart
+import com.fsck.k9.mail.testing.assertk.bodyParts
+import com.fsck.k9.mail.testing.assertk.contentTransferEncoding
+import com.fsck.k9.mail.testing.assertk.contentType
+import com.fsck.k9.mail.testing.assertk.mimeType
+import com.fsck.k9.mail.testing.assertk.parameter
+import com.fsck.k9.mail.testing.assertk.value
 import com.fsck.k9.message.MessageBuilder.Callback
 import com.fsck.k9.message.quote.InsertableHtmlContent
 import com.fsck.k9.view.RecipientSelectView
