+++ mail/common/src/test/java/com/fsck/k9/mail/internet/FlowedMessageUtilsTest.kt
@@ -2,7 +2,7 @@ package com.fsck.k9.mail.internet
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
-import com.fsck.k9.mail.crlf
+import com.fsck.k9.mail.testing.crlf
 import org.junit.Test
 
 class FlowedMessageUtilsTest {
