+++ mail/protocols/imap/src/test/java/com/fsck/k9/mail/store/imap/ImapServerSettingsValidatorTest.kt
@@ -10,12 +10,12 @@ import com.fsck.k9.mail.AuthType
 import com.fsck.k9.mail.ClientCertificateError
 import com.fsck.k9.mail.ConnectionSecurity
 import com.fsck.k9.mail.ServerSettings
-import com.fsck.k9.mail.helpers.FakeTrustManager
-import com.fsck.k9.mail.helpers.SimpleTrustedSocketFactory
 import com.fsck.k9.mail.oauth.AuthStateStorage
 import com.fsck.k9.mail.oauth.OAuth2TokenProvider
 import com.fsck.k9.mail.server.ServerSettingsValidationResult
 import com.fsck.k9.mail.store.imap.mockserver.MockImapServer
+import com.fsck.k9.mail.testing.security.FakeTrustManager
+import com.fsck.k9.mail.testing.security.SimpleTrustedSocketFactory
 import java.net.UnknownHostException
 import kotlin.test.Test
