+++ mail/protocols/pop3/src/test/java/com/fsck/k9/mail/store/pop3/Pop3FolderTest.kt
@@ -10,9 +10,9 @@ import com.fsck.k9.mail.Body
 import com.fsck.k9.mail.FetchProfile
 import com.fsck.k9.mail.MessageRetrievalListener
 import com.fsck.k9.mail.MessagingException
-import com.fsck.k9.mail.crlf
 import com.fsck.k9.mail.internet.BinaryTempFileBody
 import com.fsck.k9.mail.store.pop3.Pop3Commands.STAT_COMMAND
+import com.fsck.k9.mail.testing.crlf
 import java.io.ByteArrayOutputStream
 import java.io.File
 import java.io.IOException
