+++ mail/protocols/smtp/src/test/java/com/fsck/k9/mail/transport/smtp/SmtpTransportTest.kt
@@ -15,12 +15,12 @@ import com.fsck.k9.mail.Message
 import com.fsck.k9.mail.MessagingException
 import com.fsck.k9.mail.MissingCapabilityException
 import com.fsck.k9.mail.ServerSettings
-import com.fsck.k9.mail.XOAuth2ChallengeParserTest
 import com.fsck.k9.mail.filter.Base64
-import com.fsck.k9.mail.helpers.TestMessageBuilder
-import com.fsck.k9.mail.helpers.TestTrustedSocketFactory
 import com.fsck.k9.mail.internet.MimeMessage
 import com.fsck.k9.mail.oauth.OAuth2TokenProvider
+import com.fsck.k9.mail.testing.XOAuth2ChallengeParserTestData
+import com.fsck.k9.mail.testing.message.TestMessageBuilder
+import com.fsck.k9.mail.testing.security.TestTrustedSocketFactory
 import com.fsck.k9.mail.transport.mockServer.MockSmtpServer
 import org.junit.Test
 import org.mockito.ArgumentMatchers.anyLong
@@ -235,7 +235,7 @@ class SmtpTransportTest {
             output("250-ENHANCEDSTATUSCODES")
             output("250 AUTH XOAUTH2")
             expect("AUTH XOAUTH2 dXNlcj11c2VyAWF1dGg9QmVhcmVyIG9sZFRva2VuAQE=")
-            output("334 " + XOAuth2ChallengeParserTest.STATUS_401_RESPONSE)
+            output("334 " + XOAuth2ChallengeParserTestData.STATUS_401_RESPONSE)
             expect("")
             output("535-5.7.1 Username and Password not accepted. Learn more at")
             output("535 5.7.1 http://support.google.com/mail/bin/answer.py?answer=14257 hx9sm5317360pbc.68")
@@ -268,7 +268,7 @@ class SmtpTransportTest {
             output("250-localhost Hello client.localhost")
             output("250 AUTH XOAUTH2")
             expect("AUTH XOAUTH2 dXNlcj11c2VyAWF1dGg9QmVhcmVyIG9sZFRva2VuAQE=")
-            output("334 " + XOAuth2ChallengeParserTest.STATUS_400_RESPONSE)
+            output("334 " + XOAuth2ChallengeParserTestData.STATUS_400_RESPONSE)
             expect("")
             output("535-5.7.1 Username and Password not accepted. Learn more at")
             output("535 5.7.1 http://support.google.com/mail/bin/answer.py?answer=14257 hx9sm5317360pbc.68")
@@ -296,7 +296,7 @@ class SmtpTransportTest {
             output("250-localhost Hello client.localhost")
             output("250 AUTH XOAUTH2")
             expect("AUTH XOAUTH2 dXNlcj11c2VyAWF1dGg9QmVhcmVyIG9sZFRva2VuAQE=")
-            output("334 " + XOAuth2ChallengeParserTest.INVALID_RESPONSE)
+            output("334 " + XOAuth2ChallengeParserTestData.INVALID_RESPONSE)
             expect("")
             output("535-5.7.1 Username and Password not accepted. Learn more at")
             output("535 5.7.1 http://support.google.com/mail/bin/answer.py?answer=14257 hx9sm5317360pbc.68")
@@ -324,7 +324,7 @@ class SmtpTransportTest {
             output("250-localhost Hello client.localhost")
             output("250 AUTH XOAUTH2")
             expect("AUTH XOAUTH2 dXNlcj11c2VyAWF1dGg9QmVhcmVyIG9sZFRva2VuAQE=")
-            output("334 " + XOAuth2ChallengeParserTest.MISSING_STATUS_RESPONSE)
+            output("334 " + XOAuth2ChallengeParserTestData.MISSING_STATUS_RESPONSE)
             expect("")
             output("535-5.7.1 Username and Password not accepted. Learn more at")
             output("535 5.7.1 http://support.google.com/mail/bin/answer.py?answer=14257 hx9sm5317360pbc.68")
@@ -353,12 +353,12 @@ class SmtpTransportTest {
             output("250-ENHANCEDSTATUSCODES")
             output("250 AUTH XOAUTH2")
             expect("AUTH XOAUTH2 dXNlcj11c2VyAWF1dGg9QmVhcmVyIG9sZFRva2VuAQE=")
-            output("334 " + XOAuth2ChallengeParserTest.STATUS_400_RESPONSE)
+            output("334 " + XOAuth2ChallengeParserTestData.STATUS_400_RESPONSE)
             expect("")
             output("535-5.7.1 Username and Password not accepted. Learn more at")
             output("535 5.7.1 http://support.google.com/mail/bin/answer.py?answer=14257 hx9sm5317360pbc.68")
             expect("AUTH XOAUTH2 dXNlcj11c2VyAWF1dGg9QmVhcmVyIG5ld1Rva2VuAQE=")
-            output("334 " + XOAuth2ChallengeParserTest.STATUS_400_RESPONSE)
+            output("334 " + XOAuth2ChallengeParserTestData.STATUS_400_RESPONSE)
             expect("")
             output("535-5.7.1 Username and Password not accepted. Learn more at")
             output("535 5.7.1 http://support.google.com/mail/bin/answer.py?answer=14257 hx9sm5317360pbc.68")
@@ -490,7 +490,7 @@ class SmtpTransportTest {
         server.output("250-ENHANCEDSTATUSCODES")
         server.output("250 AUTH XOAUTH2")
         server.expect("AUTH XOAUTH2 dXNlcj11c2VyAWF1dGg9QmVhcmVyIG9sZFRva2VuAQE=")
-        server.output("334 " + XOAuth2ChallengeParserTest.STATUS_401_RESPONSE)
+        server.output("334 " + XOAuth2ChallengeParserTestData.STATUS_401_RESPONSE)
         server.expect("")
         server.output("535-5.7.1 Username and Password not accepted. Learn more at")
         server.output("535 5.7.1 http://support.google.com/mail/bin/answer.py?answer=14257 hx9sm5317360pbc.68")
