+++ mail/testing/src/main/java/com/fsck/k9/mail/testing/XOAuth2ChallengeParserTestData.java
@@ -1,9 +1,9 @@
-package com.fsck.k9.mail;
+package com.fsck.k9.mail.testing;
 
 import com.fsck.k9.mail.filter.Base64;
 
 
-public class XOAuth2ChallengeParserTest {
+public class XOAuth2ChallengeParserTestData {
     public static final String STATUS_400_RESPONSE = Base64.encode(
             "{\"status\":\"400\",\"schemes\":\"bearer mac\",\"scope\":\"https://mail.google.com/\"}");
     public static final String STATUS_401_RESPONSE = Base64.encode(
