+++ mail/testing/src/main/java/com/fsck/k9/mail/testing/security/KeyStoreProvider.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.mail.helpers;
+package com.fsck.k9.mail.testing.security;
 
 
 import java.io.InputStream;
